+++ src/main/java/org/mju_likelion/festival/auth/service/AuthService.java
@@ -10,8 +10,8 @@ import lombok.RequiredArgsConstructor;
 import org.mju_likelion.festival.auth.dto.request.UserLoginRequest;
 import org.mju_likelion.festival.auth.dto.response.UserLoginResponse;
 import org.mju_likelion.festival.auth.util.jwt.AuthenticationRole;
-import org.mju_likelion.festival.auth.util.key.RsaKeyStrategy;
-import org.mju_likelion.festival.auth.util.key.manager.RsaKeyManager;
+import org.mju_likelion.festival.auth.util.rsa_key.RsaKeyStrategy;
+import org.mju_likelion.festival.auth.util.rsa_key.manager.RsaKeyManager;
 import org.mju_likelion.festival.common.exception.BadRequestException;
 import org.mju_likelion.festival.common.exception.type.ErrorType;
 import org.mju_likelion.festival.term.domain.Term;
