+++ src/main/java/org/mju_likelion/festival/auth/service/AuthServiceUtil.java
@@ -7,9 +7,9 @@ import org.mju_likelion.festival.admin.domain.repository.AdminJpaRepository;
 import org.mju_likelion.festival.auth.util.jwt.AuthenticationRole;
 import org.mju_likelion.festival.auth.util.jwt.JwtUtil;
 import org.mju_likelion.festival.auth.util.jwt.Payload;
-import org.mju_likelion.festival.auth.util.key.RsaKeyStrategy;
-import org.mju_likelion.festival.auth.util.key.manager.RsaKeyManager;
-import org.mju_likelion.festival.auth.util.key.manager.RsaKeyManagerContext;
+import org.mju_likelion.festival.auth.util.rsa_key.RsaKeyStrategy;
+import org.mju_likelion.festival.auth.util.rsa_key.manager.RsaKeyManager;
+import org.mju_likelion.festival.auth.util.rsa_key.manager.RsaKeyManagerContext;
 import org.mju_likelion.festival.common.exception.UnauthorizedException;
 import org.mju_likelion.festival.common.exception.type.ErrorType;
 import org.mju_likelion.festival.common.util.api.mju.MjuApiUtil;
