+++ src/main/java/org/mju_likelion/festival/auth/util/rsa_key/RsaKey.java
@@ -1,4 +1,4 @@
-package org.mju_likelion.festival.auth.util.key;
+package org.mju_likelion.festival.auth.util.rsa_key;
 
 /**
  * RSA Key 정보 클래스
