+++ src/main/java/org/mju_likelion/festival/auth/util/rsa_key/manager/RsaKeyManagerContext.java
@@ -1,10 +1,10 @@
-package org.mju_likelion.festival.auth.util.key.manager;
+package org.mju_likelion.festival.auth.util.rsa_key.manager;
 
 import java.util.List;
 import java.util.Map;
 import java.util.function.Function;
 import java.util.stream.Collectors;
-import org.mju_likelion.festival.auth.util.key.RsaKeyStrategy;
+import org.mju_likelion.festival.auth.util.rsa_key.RsaKeyStrategy;
 import org.mju_likelion.festival.common.util.redis.RedisAvailabilityChecker;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
