+++ src/main/java/org/mju_likelion/festival/auth/util/rsa_key/manager/TokenRsaKeyManager.java
@@ -1,9 +1,9 @@
-package org.mju_likelion.festival.auth.util.key.manager;
+package org.mju_likelion.festival.auth.util.rsa_key.manager;
 
 import lombok.AllArgsConstructor;
-import org.mju_likelion.festival.auth.util.key.RsaKey;
-import org.mju_likelion.festival.auth.util.key.RsaKeyStrategy;
-import org.mju_likelion.festival.auth.util.key.RsaKeyUtil;
+import org.mju_likelion.festival.auth.util.rsa_key.RsaKey;
+import org.mju_likelion.festival.auth.util.rsa_key.RsaKeyStrategy;
+import org.mju_likelion.festival.auth.util.rsa_key.RsaKeyUtil;
 import org.mju_likelion.festival.common.util.token.TokenUtil;
 import org.springframework.stereotype.Service;
